requirements.txt
streamlit
openai
chromadb
pysqlite3-binary
PyPDF2
ProtoBuf==3.20